backend/pyproject.toml
[project]
name = "eduagent-backend"
version = "0.1.0"
description = "EduAgent AI — Backend API"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # LLM
    "anthropic>=0.40.0",

    # Agent / Orchestration
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",

    # RAG
    "llama-index>=0.11.0",
    "llama-index-vector-stores-postgres>=0.2.0",
    "llama-parse>=0.5.0",

    # Embeddings
    "cohere>=5.0.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",

    # Supabase (DB auth + Storage)
    "supabase>=2.3.0",

    # Cache (Upstash-compatible redis-py)
    "redis[hiredis]>=5.0.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=42.0.0",

    # HTTP client
    "httpx>=0.27.0",

    # Validation / Settings
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # Document parsing
    "python-docx>=1.1.0",

    # Video / Audio
    "yt-dlp>=2024.1.0",
    "openai-whisper>=20231117",

    # Rate limiting
    "slowapi>=0.1.9",

    # Logging
    "structlog>=24.1.0",

    # Utilities
    "python-dateutil>=2.9.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "alembic/*"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
